psi: resolved receiver calls run ordinary result-use, arity, and bounds validation
Receiver calls on parameters, locals, nested paths, type names, and
boundary parameters bypassed validate_result_use, arity checks, and
type-parameter-bound validation because the name ladder missed them.
Add a resolved-target rung (machine_state_by_symbol) plus a
boundary-trait-signature fallback so those calls rejoin ordinary
validation; thread self_is_argument so static calls count self as
arg 0; split the bodyless-boundary execution fence from
instantiation-bound checking for the resolved rung. Existing fixtures
that silently discarded non-unit results gain explicit discards.

+++ omega-rust/psi/pipeline/checked-trees-to-lowered-psi/tests/field_rhs_computation_source.rs
@@ -285,7 +285,7 @@ fn backedge_field_rhs_calls_read_each_previous_replacement() {
             transition { _ -> step() }
             state step(&mut self) {
                 self.value = identity(self.value + 1) as u16 in Wrapping;
-                Trace::observe(self.value);
+                Trace::observe(self.value as u16);
                 transition self.value < 3 { true -> step() _ -> done() }
             }
             state done(&mut self) {}
